null check

+++ modular_bandastation/balance/code/regenerative_core.dm
@@ -1,3 +1,5 @@
 /obj/item/organ/monster_core/regenerative_core/on_triggered_internal()
+	if(!owner)
+		return
 	apply_to(owner)
 	qdel(src)
